Started using Long type for ID in all places

+++ src/main/java/ua/lviv/iot/greenhouse/controllers/SensorDataController.java
@@ -121,7 +121,7 @@ public class SensorDataController {
     }
 
     @PutMapping("{id}")
-    public SensorData updateDataById(final @RequestBody SensorData sensorData, final @PathVariable("id") int id) {
+    public SensorData updateDataById(final @RequestBody SensorData sensorData, final @PathVariable("id") Long id) {
         return sensorDataService.updateDataById(id, sensorData.getData());
     }
 }
